Breaking up js file for easier troubleshooting
