+++ filters.js
@@ -1,0 +1,8 @@
+module.exports = {
+    add: {
+        upcoming: {
+            label: 'Upcoming',
+            def: true
+        }
+    }
+}
