+++ index.js
@@ -2,6 +2,9 @@ const fs = require('fs')
 const path = require('path')
 const dayjs = require('dayjs')
 
+const filters = require('./filters')
+const queries = require('./queries')
+
 module.exports = {
   extend: '@apostrophecms/piece-type',
   bundle: {
@@ -12,14 +15,6 @@ module.exports = {
     label: 'Event',
     pluralLabel: 'Events',
     sort: { start: 1 },
-  },
-  filters: {
-    add: {
-      upcoming: {
-        label: 'Upcoming',
-        def: true
-      }
-    }
   },
   columns: {
     add: {
@@ -261,45 +256,8 @@ module.exports = {
       }
     }
   },
-  queries(self, query) {
-    return {
-      builders: {
-        upcoming: {
-          async finalize() {
-            // Navigation by year, month or day should
-            // trump this filter allowing you to
-            // browse the past
-
-            const upcoming = query.get('upcoming')
-
-            if (upcoming === null) {
-              return
-            }
-
-            if (upcoming) {
-              query.and({
-                end: { $gt: new Date() }
-              })
-            } else {
-              query.and({
-                end: { $lte: new Date() }
-              })
-            }
-          },
-          launder(value) {
-            return self.apos.launder.booleanOrNull(value)
-          },
-          choices() {
-            return [
-              { value: null, label: 'Both' },
-              { value: true, label: 'Upcoming' },
-              { value: false, label: 'Past' }
-            ]
-          }
-        }
-      }
-    }
-  }
+  filters,
+  queries
 }
 
 function getBundleModuleNames() {
